First attempt to save the MMUser
1. Added some attributes and an onPause() and an onResume() method to
MainActicity, needed to save and retrieve the MMUser object.
2. Added a Supporting class and moved the getUser()-method from the
SettingsSaveButtonFragment class to it, making it possible to use it
from multiple classes.
3. Added some unimportant author information

TODO: Something does not work as intended. Could be that each activity
has its own SharedPreferences, so the data saved in the SettingsActivity
could be not accessible from the MainActivity. Not enough tested yet.