Improved Inter Process Communication (IPC)
1. The Activitys now exchange data with Extras instead of using static
methods.
2. The SettingsActivity is now called as subActivity of the
MainActivity, which improves data exchange. It fixes another problem,
which caused that you could open a MainActivity, aSettingsActivity, then
a MainActivity again and so on, and then had to go all the steps
backwards again, when clicking the back button.

Note: The MMUser object 'myself' still is not saved properly.